Possibly getting close to Alpha 2 state
Math game does not crash, will not divide by 0 or negative numbers, and gives score after 5 questions. Logic game is included even though it illogically just redirects to the list of games and has no score. Most of the screens have the same style being used. I still need to do some work on score sheets for Fast tap and Simon and also need to work on the profile.